Refactor Tradfri switch device
