+++ homeassistant/components/tradfri/base_class.py
@@ -1,0 +1,94 @@
+"""Base class for IKEA TRADFRI."""
+from pytradfri.error import PytradfriError
+
+from homeassistant.core import callback
+from . import DOMAIN as TRADFRI_DOMAIN
+
+
+class TradfriBaseDevice:
+    """GBase class for a TRADFRI device."""
+
+    def __init__(self, device, api, gateway_id):
+        """Initialize a switch."""
+        self._api = api
+        self._decice = None
+        self._device_control = None
+        self._device_data = None
+        self._name = None
+        self._available = True
+        self._gateway_id = gateway_id
+
+        self._refresh(device)
+
+    def _refresh(self, device):
+        """Refresh the switch data."""
+        self._switch = device
+
+        # Caching of switchControl and switch object
+        self._available = device.reachable
+        self._socket_control = device.device_control
+        self._switch_data = device.device_control.sockets[0]
+        self._name = device.name
+
+    @property
+    def available(self):
+        """Return True if entity is available."""
+        return self._available
+
+    async def async_added_to_hass(self):
+        """Start thread when added to hass."""
+        self._async_start_observe()
+
+    @property
+    def unique_id(self):
+        """Return unique ID for switch."""
+        return self._unique_id
+
+    @property
+    def device_info(self):
+        """Return the device info."""
+        info = self._device.device_info
+
+        return {
+            "identifiers": {(TRADFRI_DOMAIN, self._device.id)},
+            "name": self._name,
+            "manufacturer": info.manufacturer,
+            "model": info.model_number,
+            "sw_version": info.firmware_version,
+            "via_device": (TRADFRI_DOMAIN, self._gateway_id),
+        }
+
+    @property
+    def name(self):
+        """Return the display name of this switch."""
+        return self._name
+
+    @property
+    def should_poll(self):
+        """No polling needed for tradfri switch."""
+        return False
+
+    @callback
+    def _async_start_observe(self, exc=None):
+        """Start observation of switch."""
+        if exc:
+            self._available = False
+            self.async_schedule_update_ha_state()
+            _LOGGER.warning("Observation failed for %s", self._name, exc_info=exc)
+
+        try:
+            cmd = self._switch.observe(
+                callback=self._observe_update,
+                err_callback=self._async_start_observe,
+                duration=0,
+            )
+            self.hass.async_create_task(self._api(cmd))
+        except PytradfriError as err:
+            _LOGGER.warning("Observation failed, trying again", exc_info=err)
+            self._async_start_observe()
+
+    @callback
+    def _observe_update(self, device):
+        """Receive new state data for this switch."""
+        self._refresh(device)
+        self.async_schedule_update_ha_state()
